requirements.txt
kafka-python
celery
celery[kafka]
confluent-kafka
kq
requests
icecream